+++ cli/ops/worker_host.rs
@@ -4,27 +4,16 @@ use crate::deno_error::DenoError;
 use crate::deno_error::ErrorKind;
 use crate::deno_error::GetErrorKind;
 use crate::fmt_errors::JSError;
-use crate::futures::SinkExt;
-use crate::global_state::GlobalState;
 use crate::ops::json_op;
-use crate::permissions::DenoPermissions;
-use crate::startup_data;
 use crate::state::State;
-use crate::tokio_util::create_basic_runtime;
-use crate::web_worker::WebWorker;
-use crate::worker::Worker;
+use crate::web_worker;
 use crate::worker::WorkerEvent;
-use crate::worker::WorkerHandle;
 use deno_core::*;
 use futures;
-use futures::future::poll_fn;
 use futures::future::FutureExt;
 use futures::future::TryFutureExt;
-use futures::stream::StreamExt;
 use std;
 use std::convert::From;
-use std::task::Poll;
-use std::thread::JoinHandle;
 
 pub fn init(i: &mut Isolate, s: &State) {
   i.register_op(
@@ -43,158 +32,6 @@ pub fn init(i: &mut Isolate, s: &State) {
     "host_get_message",
     s.core_op(json_op(s.stateful_op(op_host_get_message))),
   );
-}
-
-fn create_web_worker(
-  name: String,
-  global_state: GlobalState,
-  permissions: DenoPermissions,
-  specifier: ModuleSpecifier,
-) -> Result<WebWorker, ErrBox> {
-  let state =
-    State::new_for_worker(global_state, Some(permissions), specifier)?;
-
-  let mut worker =
-    WebWorker::new(name.to_string(), startup_data::deno_isolate_init(), state);
-  let script = format!("bootstrapWorkerRuntime(\"{}\")", name);
-  worker.execute(&script)?;
-
-  Ok(worker)
-}
-
-// TODO(bartlomieju): this function should probably live in `cli/web_worker.rs`
-pub fn run_worker_loop(
-  rt: &mut tokio::runtime::Runtime,
-  worker: &mut Worker,
-) -> Result<(), ErrBox> {
-  let mut worker_is_ready = false;
-
-  let fut = poll_fn(|cx| -> Poll<Result<(), ErrBox>> {
-    if !worker_is_ready {
-      match worker.poll_unpin(cx) {
-        Poll::Ready(r) => {
-          if let Err(e) = r {
-            let mut sender = worker.internal_channels.sender.clone();
-            futures::executor::block_on(sender.send(WorkerEvent::Error(e)))
-              .expect("Failed to post message to host");
-          }
-          worker_is_ready = true;
-        }
-        Poll::Pending => {}
-      }
-    }
-
-    let maybe_msg = {
-      match worker.internal_channels.receiver.poll_next_unpin(cx) {
-        Poll::Ready(r) => match r {
-          Some(msg) => {
-            let msg_str = String::from_utf8(msg.to_vec()).unwrap();
-            debug!("received message from host: {}", msg_str);
-            Some(msg_str)
-          }
-          None => {
-            debug!("channel closed by host, worker event loop shuts down");
-            return Poll::Ready(Ok(()));
-          }
-        },
-        Poll::Pending => None,
-      }
-    };
-
-    if let Some(msg) = maybe_msg {
-      // TODO: just add second value and then bind using rusty_v8
-      // to get structured clone/transfer working
-      let script = format!("workerMessageRecvCallback({})", msg);
-      worker
-        .execute(&script)
-        .expect("Failed to execute message cb");
-      // Let worker be polled again
-      worker_is_ready = false;
-      worker.waker.wake();
-    }
-
-    Poll::Pending
-  });
-
-  rt.block_on(fut)
-}
-
-// TODO(bartlomieju): this function should probably live in `cli/web_worker.rs`
-// TODO(bartlomieju): check if order of actions is aligned to Worker spec
-fn run_worker_thread(
-  name: String,
-  global_state: GlobalState,
-  permissions: DenoPermissions,
-  specifier: ModuleSpecifier,
-  has_source_code: bool,
-  source_code: String,
-) -> Result<(JoinHandle<()>, WorkerHandle), ErrBox> {
-  let (handle_sender, handle_receiver) =
-    std::sync::mpsc::sync_channel::<Result<WorkerHandle, ErrBox>>(1);
-
-  let builder =
-    std::thread::Builder::new().name(format!("deno-worker-{}", name));
-  let join_handle = builder.spawn(move || {
-    // Any error inside this block is terminal:
-    // - JS worker is useless - meaning it throws an exception and can't do anything else,
-    //  all action done upon it should be noops
-    // - newly spawned thread exits
-    let result =
-      create_web_worker(name, global_state, permissions, specifier.clone());
-
-    if let Err(err) = result {
-      handle_sender.send(Err(err)).unwrap();
-      return;
-    }
-
-    let mut worker = result.unwrap();
-    // Send thread safe handle to newly created worker to host thread
-    handle_sender.send(Ok(worker.thread_safe_handle())).unwrap();
-    drop(handle_sender);
-
-    // At this point the only method of communication with host
-    // is using `worker.internal_channels`.
-    //
-    // Host can already push messages and interact with worker.
-    //
-    // Next steps:
-    // - create tokio runtime
-    // - load provided module or code
-    // - start driving worker's event loop
-
-    let mut rt = create_basic_runtime();
-
-    // TODO: run with using select with terminate
-
-    // Execute provided source code immediately
-    let result = if has_source_code {
-      worker.execute(&source_code)
-    } else {
-      // TODO(bartlomieju): add "type": "classic", ie. ability to load
-      // script instead of module
-      let load_future = worker
-        .execute_mod_async(&specifier, None, false)
-        .boxed_local();
-
-      rt.block_on(load_future)
-    };
-
-    if let Err(e) = result {
-      let mut sender = worker.internal_channels.sender.clone();
-      futures::executor::block_on(sender.send(WorkerEvent::Error(e)))
-        .expect("Failed to post message to host");
-
-      // Failure to execute script is a terminal error, bye, bye.
-      return;
-    }
-
-    run_worker_loop(&mut rt, &mut worker)
-      .expect("Unexpected error in worker loop");
-  })?;
-
-  let worker_handle = handle_receiver.recv().unwrap()?;
-
-  Ok((join_handle, worker_handle))
 }
 
 #[derive(Deserialize)]
@@ -232,7 +69,7 @@ fn op_create_worker(
     format!("USER-WORKER-{}", specifier)
   });
 
-  let (join_handle, worker_handle) = run_worker_thread(
+  let (join_handle, worker_handle) = web_worker::run_in_thread(
     worker_name,
     global_state,
     permissions,
