+++ cli/web_worker.rs
@@ -1,17 +1,22 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
+use crate::futures::SinkExt;
+use crate::global_state::GlobalState;
 use crate::ops;
+use crate::permissions::DenoPermissions;
+use crate::startup_data;
 use crate::state::State;
+use crate::tokio_util;
 use crate::worker::Worker;
+use crate::worker::WorkerEvent;
+use crate::worker::WorkerHandle;
 use deno_core;
 use deno_core::ErrBox;
+use deno_core::ModuleSpecifier;
 use deno_core::StartupData;
 use futures::future::FutureExt;
-use std::future::Future;
 use std::ops::Deref;
 use std::ops::DerefMut;
-use std::pin::Pin;
-use std::task::Context;
-use std::task::Poll;
+use std::thread::JoinHandle;
 
 /// This worker is implementation of `Worker` Web API
 ///
@@ -53,19 +58,103 @@ impl DerefMut for WebWorker {
   }
 }
 
-impl Future for WebWorker {
-  type Output = Result<(), ErrBox>;
+fn create_web_worker(
+  name: String,
+  global_state: GlobalState,
+  permissions: DenoPermissions,
+  specifier: ModuleSpecifier,
+) -> Result<WebWorker, ErrBox> {
+  let state =
+    State::new_for_worker(global_state, Some(permissions), specifier)?;
 
-  fn poll(self: Pin<&mut Self>, cx: &mut Context) -> Poll<Self::Output> {
-    let inner = self.get_mut();
-    inner.0.poll_unpin(cx)
-  }
+  let mut worker =
+    WebWorker::new(name.to_string(), startup_data::deno_isolate_init(), state);
+  let script = format!("bootstrapWorkerRuntime(\"{}\")", name);
+  worker.execute(&script)?;
+
+  Ok(worker)
+}
+
+// TODO(bartlomieju): check if order of actions is aligned to Worker spec
+pub fn run_in_thread(
+  name: String,
+  global_state: GlobalState,
+  permissions: DenoPermissions,
+  specifier: ModuleSpecifier,
+  has_source_code: bool,
+  source_code: String,
+) -> Result<(JoinHandle<()>, WorkerHandle), ErrBox> {
+  let (handle_sender, handle_receiver) =
+    std::sync::mpsc::sync_channel::<Result<WorkerHandle, ErrBox>>(1);
+
+  let builder =
+    std::thread::Builder::new().name(format!("deno-worker-{}", name));
+  let join_handle = builder.spawn(move || {
+    // Any error inside this block is terminal:
+    // - JS worker is useless - meaning it throws an exception and can't do anything else,
+    //  all action done upon it should be noops
+    // - newly spawned thread exits
+    let result =
+      create_web_worker(name, global_state, permissions, specifier.clone());
+
+    if let Err(err) = result {
+      handle_sender.send(Err(err)).unwrap();
+      return;
+    }
+
+    let mut worker = result.unwrap();
+    // Send thread safe handle to newly created worker to host thread
+    handle_sender.send(Ok(worker.thread_safe_handle())).unwrap();
+    drop(handle_sender);
+
+    // At this point the only method of communication with host
+    // is using `worker.internal_channels`.
+    //
+    // Host can already push messages and interact with worker.
+    //
+    // Next steps:
+    // - create tokio runtime
+    // - load provided module or code
+    // - start driving worker's event loop
+
+    let mut rt = tokio_util::create_basic_runtime();
+
+    // TODO: run using select with terminate
+
+    // Execute provided source code immediately
+    let result = if has_source_code {
+      worker.execute(&source_code)
+    } else {
+      // TODO(bartlomieju): add "type": "classic", ie. ability to load
+      // script instead of module
+      let load_future = worker
+        .execute_mod_async(&specifier, None, false)
+        .boxed_local();
+
+      rt.block_on(load_future)
+    };
+
+    if let Err(e) = result {
+      let mut sender = worker.internal_channels.sender.clone();
+      futures::executor::block_on(sender.send(WorkerEvent::Error(e)))
+        .expect("Failed to post message to host");
+
+      // Failure to execute script is a terminal error, bye, bye.
+      return;
+    }
+
+    rt.block_on(worker.run_event_loop())
+      .expect("Unexpected error in worker loop");
+  })?;
+
+  let worker_handle = handle_receiver.recv().unwrap()?;
+
+  Ok((join_handle, worker_handle))
 }
 
 #[cfg(test)]
 mod tests {
   use super::*;
-  use crate::ops::worker_host::run_worker_loop;
   use crate::startup_data;
   use crate::state::State;
   use crate::tokio_util;
@@ -105,7 +194,7 @@ mod tests {
       let handle = worker.thread_safe_handle();
       handle_sender.send(handle).unwrap();
       let mut rt = tokio_util::create_basic_runtime();
-      let r = run_worker_loop(&mut rt, &mut worker);
+      let r = rt.block_on(worker.run_event_loop());
       assert!(r.is_ok())
     });
 
@@ -155,7 +244,7 @@ mod tests {
       let handle = worker.thread_safe_handle();
       handle_sender.send(handle).unwrap();
       let mut rt = tokio_util::create_basic_runtime();
-      let r = run_worker_loop(&mut rt, &mut worker);
+      let r = rt.block_on(worker.run_event_loop());
       assert!(r.is_ok())
     });
